+++ frontend/src/features/org/index.ts
@@ -1,3 +1,4 @@
 import("./org-status-banner");
+import("./org-active-crawls-status");
 import("./org-uploads-dialog");
 import("./usage-history-table");
