+++ frontend/src/index.ts
@@ -3,17 +3,11 @@ import "./global";
 
 import { provide } from "@lit/context";
 import { localized, msg, str } from "@lit/localize";
-import type {
-  SlDialog,
-  SlDrawer,
-  SlSelectEvent,
-} from "@shoelace-style/shoelace";
-import clsx from "clsx";
+import type { SlDialog, SlDrawer } from "@shoelace-style/shoelace";
 import { html, nothing, type TemplateResult } from "lit";
 import { customElement, property, query, state } from "lit/decorators.js";
 import { ifDefined } from "lit/directives/if-defined.js";
 import { until } from "lit/directives/until.js";
-import { when } from "lit/directives/when.js";
 import isEqual from "lodash/fp/isEqual";
 
 import "./components";
@@ -27,7 +21,6 @@ import { notificationsContextKey } from "./context/notifications/types";
 import { viewStateContext } from "./context/view-state";
 import type { BtrixUserGuideShowEvent } from "./events/btrix-user-guide-show";
 import { OrgTab, RouteNamespace } from "./routes";
-import { ORG_NAME_MAX_LENGTH } from "./types/org";
 import type { UserInfo, UserOrg } from "./types/user";
 import { pageView, type AnalyticsTrackProps } from "./utils/analytics";
 import { type ViewState } from "./utils/APIRouter";
@@ -36,22 +29,18 @@ import AuthService, {
   type LoggedInEventDetail,
   type NeedLoginEventDetail,
 } from "./utils/AuthService";
-import { tw } from "./utils/tailwind";
 
 import { BtrixElement } from "@/classes/BtrixElement";
 import type { NavigateEventDetail } from "@/controllers/navigate";
 import { type Auth } from "@/types/auth";
-import {
-  translatedLocales,
-  type TranslatedLocaleEnum,
-} from "@/types/localization";
 import { type AppSettings } from "@/utils/app";
 import { DEFAULT_MAX_SCALE } from "@/utils/crawler";
 import localize from "@/utils/localize";
 import router, { urlForName } from "@/utils/router";
 import { AppStateService } from "@/utils/state";
 import { formatAPIUser } from "@/utils/user";
-import brandLockupColor from "~assets/brand/browsertrix-lockup-color.svg";
+
+import "@/components/layout/app-bar";
 
 type DialogContent = {
   label?: TemplateResult | string;
@@ -345,7 +334,13 @@ export class App extends BtrixElement {
   render() {
     return html`
       <div class="min-w-screen relative flex min-h-screen flex-col">
-        ${this.renderSuperadminBanner()} ${this.renderNavBar()}
+        ${this.renderSuperadminBanner()}
+        <btrix-app-bar
+          .docsUrl=${this.docsUrl}
+          .viewState=${this.viewState}
+          .orgSlugInPath=${this.orgSlugInPath}
+          @btrix-update-active-crawls-count=${this.onUpdateActiveCrawlsCount}
+        ></btrix-app-bar>
         ${this.renderAlertBanner()}
         <main
           class="relative flex flex-auto transition-[padding] md:min-h-[calc(100vh-3.125rem)]"
@@ -440,331 +435,6 @@ export class App extends BtrixElement {
                 )}
           </sl-alert>
         </div>
-      </div>
-    `;
-  }
-
-  private renderNavBar() {
-    const isSuperAdmin = this.authState && this.userInfo?.isSuperAdmin;
-
-    const showFullLogo =
-      this.viewState.route === "login" || !this.authService.authState;
-
-    return html`
-      <div class="border-b bg-neutral-50">
-        <nav
-          class="box-border flex min-h-12 flex-wrap items-center gap-2 p-3 leading-none md:py-0 xl:pl-6"
-        >
-          <div class="order-1 flex flex-1 items-center">
-            <a
-              class="items-between flex gap-2"
-              aria-label="home"
-              href=${this.homePath}
-              @click=${(e: MouseEvent) => {
-                if (isSuperAdmin) {
-                  this.clearSelectedOrg();
-                }
-                this.navigate.link(e);
-              }}
-            >
-              <div
-                class="${showFullLogo
-                  ? "w-[10.5rem]"
-                  : "w-6 lg:w-[10.5rem]"} h-6 bg-cover bg-no-repeat"
-                style="background-image: url(${brandLockupColor})"
-                role="img"
-                title="Browsertrix logo"
-              ></div>
-
-              ${isSuperAdmin
-                ? html`<btrix-badge size="large" variant="orange" outline>
-                    ${msg("Admin")}
-                  </btrix-badge>`
-                : nothing}
-            </a>
-            ${when(
-              this.userInfo,
-              () => html`
-                ${this.renderOrgs()}
-                ${isSuperAdmin && this.org?.note
-                  ? html`
-                      <btrix-popover>
-                        <span slot="content" class="whitespace-pre-line"
-                          >${this.org.note}</span
-                        >
-                        <sl-icon
-                          name="chat-left-text"
-                          class="relative flex-shrink-0 text-neutral-500"
-                          title=${msg("This org has a note")}
-                        ></sl-icon>
-                      </btrix-popover>
-                    `
-                  : nothing}
-              `,
-            )}
-          </div>
-          <div
-            class="${this.authState
-              ? "gap-4"
-              : "gap-2"} order-2 flex flex-grow-0 items-center md:order-3"
-          >
-            ${this.authState
-              ? html`${this.userInfo && !isSuperAdmin
-                    ? html`
-                        <button
-                          class="flex items-center gap-2 leading-none text-neutral-500 hover:text-primary"
-                          @click=${() => void this.showUserGuide()}
-                        >
-                          <sl-icon
-                            name="book"
-                            class="mt-px size-4 text-base"
-                          ></sl-icon>
-                          <span class="sr-only lg:not-sr-only"
-                            >${msg("User Guide")}</span
-                          >
-                        </button>
-                      `
-                    : nothing}
-                  <sl-dropdown
-                    class="ml-auto"
-                    placement="bottom-end"
-                    distance="4"
-                  >
-                    <div slot="trigger">
-                      ${this.renderNavButton({
-                        content: msg("Account"),
-                        iconName: "person-circle",
-                        href: urlForName("accountSettings"),
-                        className: tw`hidden md:block`,
-                        click: (e: MouseEvent) => {
-                          e.preventDefault();
-                        },
-                      })}
-                      <sl-avatar
-                        label=${msg("Open user menu")}
-                        shape="rounded"
-                        class="[--size:1.75rem] md:hidden"
-                        role="button"
-                      ></sl-avatar>
-                    </div>
-                    <sl-menu class="w-60 min-w-min max-w-full">
-                      <div class="px-7 py-2">${this.renderMenuUserInfo()}</div>
-                      <sl-divider></sl-divider>
-                      <sl-menu-item
-                        @click=${() => this.routeTo("/account/settings")}
-                      >
-                        <sl-icon slot="prefix" name="person-gear"></sl-icon>
-                        ${msg("Account Settings")}
-                      </sl-menu-item>
-                      ${this.userInfo?.isSuperAdmin
-                        ? html` <sl-menu-item
-                            @click=${() =>
-                              this.routeTo(urlForName("adminUsersInvite"))}
-                          >
-                            <sl-icon slot="prefix" name="person-plus"></sl-icon>
-                            ${msg("Invite Users")}
-                          </sl-menu-item>`
-                        : ""}
-                      <sl-divider></sl-divider>
-                      <sl-menu-item @click="${this.onLogOut}">
-                        <sl-icon slot="prefix" name="door-open"></sl-icon>
-                        ${msg("Log Out")}
-                      </sl-menu-item>
-                    </sl-menu>
-                  </sl-dropdown>`
-              : html`
-                  ${this.viewState.route !== "login"
-                    ? html`
-                        <sl-button
-                          size="small"
-                          variant="primary"
-                          href="/log-in"
-                          @click=${this.navigate.link}
-                        >
-                          ${msg("Sign In")}
-                        </sl-button>
-                      `
-                    : nothing}
-                  ${(translatedLocales as unknown as string[]).length > 2
-                    ? html`
-                        <btrix-user-language-select
-                          @sl-select=${this.onSelectLocale}
-                        ></btrix-user-language-select>
-                      `
-                    : nothing}
-                `}
-          </div>
-
-          ${isSuperAdmin
-            ? html`
-                <div
-                  class="order-3 flex w-full auto-cols-max items-center gap-2 md:order-2 md:w-auto"
-                >
-                  ${this.renderSuperadminNav()}
-                </div>
-              `
-            : nothing}
-        </nav>
-      </div>
-    `;
-  }
-
-  private renderNavButton({
-    content,
-    iconName,
-    href,
-    className,
-    click,
-  }: {
-    content: string | TemplateResult;
-    iconName: string;
-    href: string;
-    className?: string;
-    click?: (e: MouseEvent) => void;
-  }) {
-    const currentPage = this.viewState.pathname === href;
-
-    return html`<sl-button
-      variant="text"
-      size="small"
-      class=${clsx(
-        tw`part-[label]:font-medium part-[label]:text-neutral-700 part-[label]:hover:text-primary-600`,
-        className,
-        currentPage && tw`part-[base]:bg-primary-50`,
-      )}
-      href=${href}
-      @click=${(e: MouseEvent) => {
-        click?.(e);
-        this.navigate.link(e);
-      }}
-      aria-current=${ifDefined(currentPage ? "page" : undefined)}
-    >
-      <sl-icon slot="prefix" name=${iconName}></sl-icon>
-      ${content}
-    </sl-button>`;
-  }
-
-  private renderSuperadminNav() {
-    return html`${this.renderNavButton({
-      content: html`${msg("Active Crawls")}
-        <btrix-active-crawls-badge
-          slot="suffix"
-          class="-mt-0.5"
-          @btrix-update-active-crawls-count=${this.onUpdateActiveCrawlsCount}
-        ></btrix-active-crawls-badge>`,
-      iconName: "gear-wide-connected",
-      href: urlForName("adminCrawls"),
-    })}
-    ${this.renderNavButton({
-      content: msg("Organizations"),
-      iconName: "building-fill-gear",
-      href: urlForName("admin"),
-      click: () => {
-        this.clearSelectedOrg();
-      },
-    })}`;
-  }
-
-  private renderOrgs() {
-    const orgs = this.userInfo?.orgs;
-    if (!orgs) return;
-
-    const selectedOption = this.orgSlugInPath
-      ? orgs.find(({ slug }) => slug === this.orgSlugInPath)
-      : {
-          slug: "",
-          name: msg("All Organizations"),
-        };
-
-    if (!selectedOption) {
-      return;
-    }
-
-    // Limit org name display for orgs created before org name max length restriction
-    const orgNameLength = ORG_NAME_MAX_LENGTH;
-
-    return html`
-      <div role="separator" class="mx-2.5 h-7 w-0 border-l"></div>
-      <div class="max-w-32 truncate sm:max-w-52 md:max-w-none">
-        ${selectedOption.slug
-          ? html`
-              <a
-                class="font-medium text-neutral-600"
-                href=${`${this.navigate.orgBasePath}/${OrgTab.Dashboard}`}
-                @click=${this.navigate.link}
-              >
-                ${selectedOption.name.slice(0, orgNameLength)}
-              </a>
-            `
-          : html`
-              <span class="text-neutral-500">
-                ${selectedOption.name.slice(0, orgNameLength)}
-              </span>
-            `}
-      </div>
-      ${when(
-        orgs.length > 1,
-        () => html`
-          <sl-dropdown placement="bottom-end">
-            <sl-icon-button
-              slot="trigger"
-              name="chevron-expand"
-              label=${msg("Expand org list")}
-            ></sl-icon-button>
-            <sl-menu
-              @sl-select=${(e: CustomEvent<{ item: { value: string } }>) => {
-                const { value } = e.detail.item;
-                if (value) {
-                  this.routeTo(
-                    `/${RouteNamespace.PrivateOrgs}/${value}/${OrgTab.Dashboard}`,
-                  );
-                } else {
-                  if (this.userInfo) {
-                    this.clearSelectedOrg();
-                  }
-                  this.routeTo(`/`);
-                }
-              }}
-            >
-              ${when(
-                this.userInfo?.isSuperAdmin,
-                () => html`
-                  <sl-menu-item
-                    type="checkbox"
-                    value=""
-                    ?checked=${!selectedOption.slug}
-                    >${msg("All Organizations")}</sl-menu-item
-                  >
-                  <sl-divider></sl-divider>
-                `,
-              )}
-              ${orgs.map(
-                (org) => html`
-                  <sl-menu-item
-                    type="checkbox"
-                    value=${org.slug}
-                    ?checked=${org.slug === selectedOption.slug}
-                    >${org.name.slice(0, orgNameLength)}</sl-menu-item
-                  >
-                `,
-              )}
-            </sl-menu>
-          </sl-dropdown>
-        `,
-      )}
-    `;
-  }
-
-  private renderMenuUserInfo() {
-    if (!this.userInfo) return;
-
-    return html`
-      <div class="mb-1.5 text-xs text-neutral-500">${msg("Logged in as")}</div>
-      <div class="whitespace-nowrap font-medium text-neutral-700">
-        ${this.userInfo.email}
-        ${this.userInfo.isSuperAdmin
-          ? html`<btrix-badge>${msg("Admin")}</btrix-badge>`
-          : nothing}
       </div>
     `;
   }
@@ -1028,16 +698,10 @@ export class App extends BtrixElement {
   }
 
   private readonly onUpdateActiveCrawlsCount = (e: CustomEvent<number>) => {
+    e.stopPropagation();
+
     this.activeCrawlsCount = e.detail;
   };
-
-  onSelectLocale(e: SlSelectEvent) {
-    const locale = e.detail.item.value as TranslatedLocaleEnum;
-
-    if (locale !== this.appState.userPreferences?.language) {
-      AppStateService.partialUpdateUserPreferences({ language: locale });
-    }
-  }
 
   onLogOut(event: CustomEvent<{ redirect?: boolean } | null>) {
     const detail = event.detail || {};
@@ -1183,9 +847,5 @@ export class App extends BtrixElement {
         }
       },
     );
-  }
-
-  private clearSelectedOrg() {
-    AppStateService.updateOrgSlug(null);
   }
 }
